requirements.txt
openai>=1.0.0
elevenlabs==0.2.27
moviepy==1.0.3
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.0 
